Move db functionality into its own module
