+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod arg_parser;
 pub mod args;
+mod db;
 pub mod thin_client;
 pub mod tokens;
